[Init] Add identity initializer
This is a step towards allowing tests like A * I = A.

As is, it still doesn't do what we want, since like the others, it tried
to initialize all tensors with identity, so any non square tensor will
fail.

The next step is to add options like "--identity-b" together with
"--splat-to-random", so that only the B matrix is replaced by identity.

Also remove some unused initializers.